Fix null pointer handling in lastKey(). Add cast to double in cosineSimilarity(). Add NoSuchElementException in next().

+++ src/main/java/org/apache/commons/collections4/bloomfilter/SetOperations.java
@@ -116,7 +116,7 @@ public final class SetOperations {
         final int numerator = andCardinality(first, second);
         // Given that the cardinality is an int then the product as a double will not
         // overflow, we can use one sqrt:
-        return numerator == 0 ? 0 : numerator / Math.sqrt(cardinality(first) * cardinality(second));
+        return numerator == 0 ? 0 : numerator / Math.sqrt(cardinality(first) * (double) cardinality(second));
     }
 
     /**
@@ -136,7 +136,7 @@ public final class SetOperations {
         final int numerator = andCardinality(first, second);
         // Given that the cardinality is an int then the product as a double will not
         // overflow, we can use one sqrt:
-        return numerator == 0 ? 0 : numerator / Math.sqrt(first.cardinality() * second.cardinality());
+        return numerator == 0 ? 0 : numerator / Math.sqrt(first.cardinality() * (double) second.cardinality());
     }
 
     /**
